requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
pandas==2.1.1
numpy==1.26.2
scikit-learn==1.2.2
lightgbm==4.0.0
dvc==2.70.0
mlflow==2.6.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
alembic==1.11.1
prometheus-client==0.16.0
pytest==7.3.1
joblib==1.3.2
gunicorn==21.2.0
uvloop==0.18.0
